# Flush once per mapped over execution