--- a/lib/galaxy/tools/actions/data_manager.py
+++ b/lib/galaxy/tools/actions/data_manager.py
@@ -10,7 +10,7 @@ class DataManagerToolAction(DefaultToolAction):
 
     def execute(self, tool, trans, **kwds):
         rval = super(DataManagerToolAction, self).execute(tool, trans, **kwds)
-        if isinstance(rval, tuple) and len(rval) == 2 and isinstance(rval[0], trans.app.model.Job):
+        if isinstance(rval, tuple) and len(rval) >= 2 and isinstance(rval[0], trans.app.model.Job):
             assoc = trans.app.model.DataManagerJobAssociation(job=rval[0], data_manager_id=tool.data_manager_id)
             trans.sa_session.add(assoc)
             trans.sa_session.flush()