Groundwork in the first commit by @jmchilton, originally in https://github.com/galaxyproject/galaxy/pull/6563.
The idea is to push flushes out until all jobs have been created in-memory, which should be much more efficient.