@mvdbeek 2020-07-13
https://github.com/galaxyproject/galaxy/pull/9985/commits/ef70994113d1ba146b19f11b3979d67c1b42bcc0 is similar to https://github.com/galaxyproject/galaxy/pull/9834 and we probably don't need both. Instead of skipping hids for HDCAs I treat HDCAs and HDAs uniformly when adding them to the history, which feels a little easier to think about when creating collections.