@mvdbeek 2020-07-14
Thanks John! @bgruening @wm75 I've done some timings for fastp. Input is always a list of 1000 paired collections in a list (except without the PR, didn't want to wait for hours for it to finish ... )

Without the PR (so current dev), I canceled the test, but you can see it takes about 8 seconds per job
```
*** PROFILER RESULTS ***
execute (/Users/mvandenb/src/galaxy/lib/galaxy/tools/execute.py:33)
function called 1 times

         133939307 function calls (131052641 primitive calls) in 205.615 seconds

   Ordered by: cumulative time, internal time, call count
   List reduced from 1997 to 40 due to restriction <40>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000  205.701  205.701 execute.py:33(execute)
       25    0.000    0.000  203.349    8.134 execute.py:54(execute_single_job)
       25    0.000    0.000  203.190    8.128 __init__.py:1587(handle_single_execution)
       25    0.001    0.000  203.189    8.128 __init__.py:1675(execute)
       25    0.004    0.000  203.188    8.128 __init__.py:279(execute)
       73    0.004    0.000  199.606    2.734 __init__.py:366(handle_output)
397962/397956    0.670    0.000  196.087    0.000 attributes.py:279(__get__)
626405/331506    0.941    0.000  195.711    0.001 attributes.py:699(get)
       73    0.002    0.000  193.866    2.656 __init__.py:937(determine_output_format)
 49047/49    0.258    0.000  192.896    3.937 __init__.py:4095(dataset_elements)
   298854    1.367    0.000  191.585    0.001 strategies.py:665(_load_for_state)
   196510    0.352    0.000  186.976    0.001 <string>:1(<lambda>)
   196510    1.878    0.000  186.624    0.001 strategies.py:772(_emit_lazyload)
   147143    0.112    0.000  157.315    0.001 __init__.py:4510(is_collection)
   147143    0.190    0.000  157.203    0.001 __init__.py:4498(element_type)
   147212    3.299    0.000  144.661    0.001 baked.py:557(_load_on_pk_identity)
   196510    1.699    0.000   97.698    0.000 baked.py:421(__iter__)
   197267    0.700    0.000   88.687    0.000 query.py:3501(_execute_and_instances)
   200354    0.305    0.000   77.484    0.000 base.py:952(execute)
   200354    0.229    0.000   77.123    0.000 elements.py:296(_execute_on_connection)
   200354    1.138    0.000   76.895    0.000 base.py:1088(_execute_clauseelement)
   542325    2.005    0.000   74.481    0.000 loading.py:35(instances)
   200354    1.683    0.000   74.403    0.000 base.py:1195(_execute_context)
   200306    0.154    0.000   65.291    0.000 default.py:589(do_execute)
   200306   62.478    0.000   65.138    0.000 {method 'execute' of 'psycopg2.extensions.cursor' objects}
   197266    0.197    0.000   36.525    0.000 loading.py:59(<listcomp>)
   197266    0.818    0.000   36.328    0.000 query.py:4345(row_processor)
295243/197266    9.021    0.000   35.393    0.000 loading.py:354(_instance_processor)
    49298    0.523    0.000   32.562    0.001 baked.py:539(all)
2844250/1766472    3.051    0.000   23.998    0.000 interfaces.py:559(create_row_processor)
   197291    0.640    0.000   22.744    0.000 result.py:1268(fetchall)
   200353    0.429    0.000   20.143    0.000 result.py:926(_soft_close)
   197320    0.446    0.000   19.363    0.000 base.py:899(close)
   197319    0.284    0.000   18.916    0.000 base.py:1031(close)
   197320    0.299    0.000   18.632    0.000 base.py:858(_checkin)
   197320    0.697    0.000   18.333    0.000 base.py:671(_finalize_fairy)
97977/97973    0.585    0.000   17.719    0.000 strategies.py:1968(create_row_processor)
   197319    0.525    0.000   14.602    0.000 base.py:872(_reset)
   197319    0.681    0.000   13.995    0.000 default.py:539(do_rollback)
   197319   13.147    0.000   13.147    0.000 {method 'rollback' of 'psycopg2.extensions.connection' objects}
```

With this PR (https://github.com/galaxyproject/galaxy/pull/9985/commits/14f40c6fb14be462bf4002d1aff20befd8951453) it's at 0.141959 seconds per job:
```
    ```
    *** PROFILER RESULTS ***
    execute (/Users/mvandenb/src/galaxy/lib/galaxy/tools/execute.py:33)
    function called 1 times

             157053988 function calls (154145272 primitive calls) in 141.959 seconds

       Ordered by: cumulative time, internal time, call count
       List reduced from 2008 to 40 due to restriction <40>

       ncalls  tottime  percall  cumtime  percall filename:lineno(function)
            1    0.015    0.015  142.174  142.174 execute.py:33(execute)
         1005    0.163    0.000   88.113    0.088 session.py:2489(flush)
         1004    0.357    0.000   87.866    0.088 session.py:2542(_flush)
         1000    0.014    0.000   77.615    0.078 manager.py:52(enqueue)
         1000    0.005    0.000   73.950    0.074 handlers.py:423(assign_handler)
         1000    0.006    0.000   73.867    0.074 handlers.py:283(_assign_handler_direct)
         1000    0.012    0.000   73.843    0.074 handlers.py:463(_timed_flush_obj)
         1007    0.015    0.000   53.461    0.053 session.py:501(commit)
         1007    6.048    0.006   52.696    0.052 session.py:386(_remove_snapshot)
         1000    0.014    0.000   44.539    0.045 execute.py:54(execute_single_job)
         1000    0.009    0.000   43.977    0.044 __init__.py:1587(handle_single_execution)
         1000    0.016    0.000   43.967    0.044 __init__.py:1681(execute)
         1000    0.123    0.000   43.951    0.044 __init__.py:279(execute)
      6083134   14.954    0.000   42.490    0.000 state.py:567(_expire)
    17203069/17203063    8.582    0.000   33.137    0.000 attributes.py:279(__get__)
        56047    0.087    0.000   27.250    0.000 base.py:952(execute)
        56047    0.061    0.000   27.147    0.000 elements.py:296(_execute_on_connection)
        56047    0.326    0.000   27.086    0.000 base.py:1088(_execute_clauseelement)
         3000    0.098    0.000   25.467    0.008 __init__.py:366(handle_output)
    580205/555189    0.487    0.000   25.173    0.000 attributes.py:699(get)
         3000    0.086    0.000   22.502    0.008 __init__.py:952(determine_output_format)
    6458318/6452318    3.109    0.000   22.220    0.000 attr.py:257(__call__)
        56047    0.475    0.000   20.985    0.000 base.py:1195(_execute_context)
    2002000/2000    4.659    0.000   20.442    0.010 __init__.py:4095(dataset_elements)
         1004    0.009    0.000   20.059    0.020 unitofwork.py:402(execute)
            3    0.000    0.000   17.967    5.989 mapping.py:2827(db_next_hid)
    1007/1005    0.003    0.000   17.898    0.018 session.py:899(begin)
    1007/1005    0.005    0.000   17.896    0.018 session.py:221(__init__)
            1    0.000    0.000   17.893   17.893 __init__.py:1791(add_datasets)
    1007/1005    0.004    0.000   17.890    0.018 session.py:338(_take_snapshot)
            1    0.005    0.005   17.729   17.729 __init__.py:1811(__add_datasets_optimized)
        18031    0.070    0.000   14.605    0.001 query.py:3501(_execute_and_instances)
        56042    0.040    0.000   14.442    0.000 default.py:589(do_execute)
        56042   14.071    0.000   14.402    0.000 {method 'execute' of 'psycopg2.extensions.cursor' objects}
         7013    0.019    0.000   14.137    0.002 loading.py:190(load_on_ident)
         7013    0.060    0.000   14.118    0.002 loading.py:211(load_on_pk_identity)
         7014    0.009    0.000   14.022    0.002 query.py:3417(one)
         7014    0.146    0.000   14.014    0.002 query.py:3381(one_or_none)
         1037    0.033    0.000   14.013    0.014 persistence.py:184(save_obj)
      6006000    1.807    0.000   12.632    0.000 __init__.py:4510(is_collection)
    ```
And I have another commit that just flushes once after assigning all jobs to a handler (will put this in another PR since I haven't tested the effect it could have on other assignment methods):
    ```
    *** PROFILER RESULTS ***
    execute (/Users/mvandenb/src/galaxy/lib/galaxy/tools/execute.py:33)
    function called 1 times

             86134617 function calls (83581503 primitive calls) in 81.095 seconds

       Ordered by: cumulative time, internal time, call count
       List reduced from 1967 to 40 due to restriction <40>

       ncalls  tottime  percall  cumtime  percall filename:lineno(function)
            1    0.014    0.014   81.309   81.309 execute.py:33(execute)
         1000    0.015    0.000   46.217    0.046 execute.py:54(execute_single_job)
         1000    0.010    0.000   45.634    0.046 __init__.py:1587(handle_single_execution)
         1000    0.016    0.000   45.624    0.046 __init__.py:1681(execute)
         1000    0.127    0.000   45.608    0.046 __init__.py:279(execute)
    17201067/17201061    8.780    0.000   31.037    0.000 attributes.py:279(__get__)
            6    0.135    0.022   28.675    4.779 session.py:2489(flush)
            5    0.053    0.011   28.458    5.692 session.py:2542(_flush)
         3000    0.102    0.000   26.263    0.009 __init__.py:366(handle_output)
        54048    0.080    0.000   24.775    0.000 base.py:952(execute)
        54048    0.061    0.000   24.681    0.000 elements.py:296(_execute_on_connection)
        54048    0.320    0.000   24.620    0.000 base.py:1088(_execute_clauseelement)
         3000    0.088    0.000   23.180    0.008 __init__.py:952(determine_output_format)
    579205/554189    0.501    0.000   22.908    0.000 attributes.py:699(get)
    2002000/2000    4.790    0.000   21.080    0.011 __init__.py:4095(dataset_elements)
        54048    0.465    0.000   19.773    0.000 base.py:1195(_execute_context)
            5    0.003    0.001   19.122    3.824 unitofwork.py:402(execute)
            3    0.000    0.000   18.502    6.167 mapping.py:2827(db_next_hid)
          8/6    0.000    0.000   18.418    3.070 session.py:899(begin)
          8/6    0.000    0.000   18.418    3.070 session.py:221(__init__)
          8/6    0.000    0.000   18.418    3.070 session.py:338(_take_snapshot)
            1    0.000    0.000   18.199   18.199 __init__.py:1791(add_datasets)
            1    0.006    0.006   18.018   18.018 __init__.py:1811(__add_datasets_optimized)
           38    0.020    0.001   13.474    0.355 persistence.py:184(save_obj)
        54042    0.041    0.000   13.244    0.000 default.py:589(do_execute)
        54042   12.889    0.000   13.203    0.000 {method 'execute' of 'psycopg2.extensions.cursor' objects}
      6006000    1.842    0.000   12.992    0.000 __init__.py:4511(is_collection)
        17031    0.068    0.000   12.413    0.001 query.py:3501(_execute_and_instances)
        11012    0.026    0.000   11.658    0.001 scoping.py:162(do)
         1000    0.015    0.000   11.495    0.011 __init__.py:246(_collect_inputs)
      6006000    2.824    0.000   11.151    0.000 __init__.py:4499(element_type)
         6013    0.016    0.000   11.078    0.002 loading.py:190(load_on_ident)
           35    0.002    0.000   11.071    0.316 unitofwork.py:585(execute)
         6013    0.051    0.000   11.063    0.002 loading.py:211(load_on_pk_identity)
         6014    0.008    0.000   10.983    0.002 query.py:3417(one)
         6014    0.129    0.000   10.975    0.002 query.py:3381(one_or_none)
           38    0.354    0.009   10.877    0.286 persistence.py:1039(_emit_insert_statements)
          168    0.000    0.000   10.810    0.064 unitofwork.py:520(execute_aggregate)
       109036    0.197    0.000   10.631    0.000 strategies.py:665(_load_for_state)
        11016    0.023    0.000   10.154    0.001 <string>:1(<lambda>)

    ```
This is 81ms per job now, or a 100 fold speedup. And there are still a few more things to optimize, but this should be more than enough to keep our compute resources busy :). I think we can probably go down to ~ 60 ms per job without any larger changes.